--- serverless/redaction-digest/.env-example
@@ -1,11 +1,0 @@
-# Required
-DB_URI="postgres://localhost:5432/postmangovsg_dev"
-DB_READ_REPLICA_URI="postgres://localhost:5432/postmangovsg_dev"
-# SENTRY_DSN="https://123.ingest.sentry.io/456"
-
-# You can override other config too
-# See config.ts for the full list
-# SES_HOST=""
-# SES_PASS=""
-# SES_USER=""
-# SES_PORT=""
