--- serverless/redaction-digest/.prettierrc.js
@@ -1,4 +1,0 @@
-module.exports = {
-  semi: false,
-  singleQuote: true,
-}
