--- serverless/redaction-digest/examples/event-cloudwatch-scheduled-event.js
@@ -1,15 +1,0 @@
-/* eslint-disable */
-
-// Example JSON from CloudWatch Scheduled Events
-module.exports = {
-  "id": "cdc73f9d-aea9-11e3-9d5a-835b769c0d9c",
-  "detail-type": "Scheduled Event",
-  "source": "aws.events",
-  "account": "123456789012",
-  "time": "2020-01-01T00:00:00Z",
-  "region": "ap-southeast-1",
-  "resources": [
-    "arn:aws:events:ap-southeast-1:123456789012:rule/ExampleRule"
-  ],
-  "detail": {}
-}
