--- serverless/unsubscribe-digest/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "unsubscribe-digest",
-  "version": "1.0.0",
-  "description": "Send unsubscribe digests to campaign owners",
-  "main": "src/index.ts",
-  "scripts": {
-    "build": "rm -rf build && tsc && cp -r src/assets build/assets",
-    "dev": "npm run build && tsc-watch --onSuccess \"node -r dotenv/config build/dev\"",
-    "test": "echo 'No tests yet for unsubscribe-digest'"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/opengovsg/postmangovsg.git"
-  },
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@sentry/node": "5.30.0",
-    "aws-sdk": "2.1359.0",
-    "axios": "0.27.2",
-    "convict": "6.2.4",
-    "dotenv": "8.6.0",
-    "lodash": "4.17.21",
-    "module-alias": "2.2.2",
-    "nodemailer": "6.6.2",
-    "pg": "8.5.1",
-    "pg-connection-string": "2.4.0",
-    "reflect-metadata": "0.1.13",
-    "sequelize": "6.29.1",
-    "sequelize-typescript": "2.1.3",
-    "source-map-support": "0.5.19"
-  },
-  "devDependencies": {
-    "@types/bluebird": "3.5.32",
-    "@types/convict": "5.2.2",
-    "@types/lodash": "4.14.157",
-    "@types/node": "14.0.23",
-    "@types/nodemailer": "6.4.0",
-    "@types/validator": "13.1.0",
-    "tsc-watch": "4.6.2",
-    "typescript": "4.7.4"
-  },
-  "_moduleAliases": {
-    "@shared": "../../shared/build"
-  }
-}
