+++ sst/.nvmrc
@@ -1,1 +1,1 @@
-16
+16.15
