+++ sst/package.json
@@ -14,6 +14,7 @@
   },
   "devDependencies": {
     "@tsconfig/node16": "^16.1.0",
+    "@types/lodash": "^4.14.195",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "aws-cdk-lib": "2.84.0",
     "constructs": "10.1.156",
@@ -31,7 +32,6 @@
   ],
   "dependencies": {
     "aws-cdk": "^2.87.0",
-    "aws4-axios": "^3.2.24",
     "cronitor": "^2.3.3",
     "dotenv": "^16.3.1"
   },
