+++ sst/packages/core/package.json
@@ -8,6 +8,7 @@
   },
   "devDependencies": {
     "@types/node": "^20.4.1",
+    "@types/pg": "^8.10.2",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "eslint": "^8.44.0",
@@ -21,7 +22,13 @@
   "dependencies": {
     "axios": "^1.4.0",
     "drizzle-orm": "^0.27.2",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.4",
+    "pg": "^8.11.1",
+    "pg-hstore": "^2.3.4",
     "postgres": "^3.3.5",
+    "sequelize": "^6.29.1",
+    "sequelize-typescript": "^2.1.3",
     "zod": "^3.21.4"
   }
 }
